+++ types/dragino/ldds04/uplink.spec.js
@@ -1,0 +1,71 @@
+const chai = require("chai");
+
+const rewire = require("rewire");
+const utils = require("test-utils");
+
+const { assert } = chai;
+
+describe("Dragino LDDS04 Uplink", () => {
+  let defaultSchema = null;
+  let consume = null;
+  before((done) => {
+    const script = rewire("./uplink.js");
+    consume = utils.init(script);
+    utils
+      .loadSchema(`${__dirname}/default.schema.json`)
+      .then((parsedSchema) => {
+        defaultSchema = parsedSchema;
+        done();
+      });
+  });
+
+  let lifecycleSchema = null;
+  before((done) => {
+    utils
+      .loadSchema(`${__dirname}/lifecycle.schema.json`)
+      .then((parsedSchema) => {
+        lifecycleSchema = parsedSchema;
+        done();
+      });
+  });
+
+  describe("consume()", () => {
+    it("should decode the Dragino LDDS04 report uplink", () => {
+      const data = {
+        data: {
+          port: 2,
+          payloadHex: "0D4A03160318031A031501",
+        },
+      };
+
+      utils.expectEmits((type, value) => {
+        assert.equal(type, "sample");
+        assert.isNotNull(value);
+        assert.typeOf(value.data, "object");
+
+        assert.equal(value.topic, "default");
+        assert.equal(value.data.distance1, 79);
+        assert.equal(value.data.distance2, 79.2);
+        assert.equal(value.data.distance3, 79.4);
+        assert.equal(value.data.distance4, 78.9);
+        assert.equal(value.data.extTrigger, false);
+
+        utils.validateSchema(value.data, defaultSchema, { throwError: true });
+      });
+
+      utils.expectEmits((type, value) => {
+        assert.equal(type, "sample");
+        assert.isNotNull(value);
+        assert.typeOf(value.data, "object");
+
+        assert.equal(value.topic, "lifecycle");
+        assert.equal(value.data.batteryVoltage, 3.402);
+        assert.equal(value.data.batteryLevel, 80);
+
+        utils.validateSchema(value.data, lifecycleSchema, { throwError: true });
+      });
+
+      consume(data);
+    });
+  });
+});