+++ types/dragino/lht52/uplink.spec.js
@@ -1,0 +1,67 @@
+const chai = require("chai");
+
+const rewire = require("rewire");
+const utils = require("test-utils");
+
+const { assert } = chai;
+
+describe("Dragino LHT65 Uplink", () => {
+  let defaultSchema = null;
+  let consume = null;
+  before((done) => {
+    const script = rewire("./uplink.js");
+    consume = utils.init(script);
+    utils
+      .loadSchema(`${__dirname}/default.schema.json`)
+      .then((parsedSchema) => {
+        defaultSchema = parsedSchema;
+        done();
+      });
+  });
+
+  let externalSchema = null;
+  before((done) => {
+    utils
+      .loadSchema(`${__dirname}/external.schema.json`)
+      .then((parsedSchema) => {
+        externalSchema = parsedSchema;
+        done();
+      });
+  });
+
+  let lifecycleSchema = null;
+  before((done) => {
+    utils
+      .loadSchema(`${__dirname}/lifecycle.schema.json`)
+      .then((parsedSchema) => {
+        lifecycleSchema = parsedSchema;
+        done();
+      });
+  });
+
+  describe("consume()", () => {
+    it("should decode the Dragino LHT52 report uplink", () => {
+      const data = {
+        data: {
+          port: 2,
+          payloadHex: "08CD0220F5C60161CD4EDD",
+        },
+      };
+
+      utils.expectEmits((type, value) => {
+        assert.equal(type, "sample");
+        assert.isNotNull(value);
+        assert.typeOf(value.data, "object");
+
+        assert.equal(value.topic, "default");
+        assert.equal(value.data.humidity, 54.4);
+        assert.equal(value.data.temperature, 22.53);
+        assert.equal(value.data.extTemperature, -26.18);
+
+        utils.validateSchema(value.data, defaultSchema, { throwError: true });
+      });
+
+      consume(data);
+    });
+  });
+});