+++ types/dragino/ltc2/uplink.js
@@ -7,14 +7,14 @@ function consume(event) {
   const ext = bytes[2] & 0x0f;
   lifecycle.batteryVoltage = (((bytes[0] << 8) | bytes[1]) & 0x3fff) / 1000;
   let batteryLevel =
-    Math.round((data.batteryVoltage - 2.45) / 0.0115 / 10) * 10;
+    Math.round((lifecycle.batteryVoltage - 2.45) / 0.0115 / 10) * 10;
 
   if (batteryLevel > 100) {
     batteryLevel = 100;
   } else if (batteryLevel < 0) {
     batteryLevel = 0;
   }
-  data.batteryLevel = batteryLevel;
+  lifecycle.batteryLevel = batteryLevel;
 
   if (ext === 0x01) {
     data.channel1temp = parseFloat(
@@ -38,4 +38,7 @@ function consume(event) {
       (((bytes[5] << 8) | bytes[6]) / 100).toFixed(2),
     );
   }
+
+  emit("sample", { data, topic: "default" });
+  emit("sample", { data: lifecycle, topic: "lifecycle" });
 }