+++ types/elsys/ers-co2/uplink.spec.js
@@ -33,6 +33,7 @@ describe("Elsys CO2 uplink", () => {
     it("should decode Elsys CO2 payload", () => {
       const data = {
         data: {
+          port: 5,
           payloadHex: "0100f4022804000b0601d8070e3e",
         },
       };