+++ types/elsys/ers-desk/uplink.spec.js
@@ -43,6 +43,7 @@ describe("Elsys desk uplink", () => {
     it("should decode Elsys desk payload", () => {
       const data = {
         data: {
+          port: 5,
           payloadHex: "05011000f801041101",
         },
       };
@@ -54,6 +55,7 @@ describe("Elsys desk uplink", () => {
 
         assert.equal(value.data.motion, 1);
         assert.equal(value.data.occupancy, 1);
+        assert.equal(value.data.occupied, true);
 
         utils.validateSchema(value.data, occupancySchema, { throwError: true });
       });
@@ -64,6 +66,7 @@ describe("Elsys desk uplink", () => {
     it("should decode Elsys desk Default + Motion payload", () => {
       const data = {
         data: {
+          port: 5,
           payloadHex: "0100f102250400060505070e001100",
         },
       };
@@ -99,6 +102,7 @@ describe("Elsys desk uplink", () => {
 
         assert.equal(value.data.motion, 5);
         assert.equal(value.data.occupancy, 0);
+        assert.equal(value.data.occupancy, false);
 
         utils.validateSchema(value.data, occupancySchema, { throwError: true });
       });