+++ types/elsys/ers-eye/uplink.spec.js
@@ -43,6 +43,7 @@ describe("Elsys eye uplink", () => {
     it("should decode Elsys eye payload", () => {
       const data = {
         data: {
+          port: 5,
           payloadHex: "05011101",
         },
       };
@@ -54,6 +55,7 @@ describe("Elsys eye uplink", () => {
 
         assert.equal(value.data.motion, 1);
         assert.equal(value.data.occupancy, 1);
+        assert.equal(value.data.occupied, true);
 
         utils.validateSchema(value.data, occupancySchema, { throwError: true });
       });
@@ -64,6 +66,7 @@ describe("Elsys eye uplink", () => {
     it("should decode elsys eye default + motion payload", () => {
       const data = {
         data: {
+          port: 5,
           payloadHex: "0100e102280401a00500070dff1102",
         },
       };
@@ -99,6 +102,7 @@ describe("Elsys eye uplink", () => {
 
         assert.equal(value.data.motion, 0);
         assert.equal(value.data.occupancy, 2);
+        assert.equal(value.data.occupied, true);
 
         utils.validateSchema(value.data, occupancySchema, { throwError: true });
       });