+++ types/milesight/EM500-UDL/uplink.js
@@ -1,3 +1,34 @@
+function getFillLevel(device, distance) {
+  if (device !== undefined && distance !== undefined) {
+    if (device.customFields !== undefined) {
+      const { customFields } = device;
+      let scaleLength = null;
+      let sensorDistance = 0;
+
+      if (customFields.containerHeight !== undefined) {
+        scaleLength = Number(device.customFields.containerHeight);
+      }
+
+      if (customFields.installationOffset !== undefined) {
+        sensorDistance = Number(device.customFields.installationOffset);
+      }
+
+      if (scaleLength !== null) {
+        const percentExact =
+          (100 / scaleLength) * (scaleLength - (distance - sensorDistance));
+        let fillLevel = Math.round(percentExact);
+        if (fillLevel > 100) {
+          fillLevel = 100;
+        } else if (fillLevel < 0) {
+          fillLevel = 0;
+        }
+        return fillLevel;
+      }
+    }
+  }
+  return undefined;
+}
+
 // Parse Hex Byte Array
 function parseHexString(str) {
   const result = [];
@@ -34,7 +65,11 @@ function consume(event) {
     }
     // DISTANCE
     else if (channelId === 0x03 && channelType === 0x82) {
-      decoded.distance = readUInt16LE(bytes.slice(i, i + 2));
+      decoded.distance = readUInt16LE(Array.from(bytes).slice(i, i + 2));
+      const fillLevel = getFillLevel(event.device, decoded.distance);
+      if (fillLevel !== undefined) {
+        decoded.fillLevel = fillLevel;
+      }
       i += 2;
     } else {
       break;